Move waiting methods to new documents file, add methods to wait for deleted document
